backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
sendgrid
openai
reportlab
requests
